# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the official data model for BARAFIRI API documentation.",
        "title": "Barafiri Backend API documentation",
        "termsOfService": "http://my-gpi.io/terms/",
        "contact": {
            "name": "Endy Apinageri",
            "url": "https://www.endyapina.com",
            "email": "endy.apinageri@my-gpi.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8112",
    "basePath": "/api/v1",
    "paths": {
        "/apicore/authentication/login": {
            "post": {
                "description": "Send user data to the main authentication login endpoint endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Endpoint to send user data for login",
                "parameters": [
                    {
                        "description": "The User Login data",
                        "name": "LoginData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LoginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Registration default data",
                        "schema": {
                            "$ref": "#/definitions/controller.LoginObject"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/authentication/pasword/change": {
            "post": {
                "description": "When a user decides to change password, frontend call api with old and new password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Change password updates a user password with a current one",
                "parameters": [
                    {
                        "description": "The new password changes",
                        "name": "ChangePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ChangePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check email for password mail and link",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/authentication/pasword/forget": {
            "post": {
                "description": "When a user forgets password, forgetPassword endpoint handles sending a new link for password recovery",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Endpoint to request for new user password",
                "parameters": [
                    {
                        "description": "The user email",
                        "name": "string",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check email for password mail and link",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/authentication/registration": {
            "post": {
                "description": "Send user data to the main authentication endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Endpoint to send user data for registration",
                "parameters": [
                    {
                        "description": "The User Registration data",
                        "name": "RegistrationData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.RegistrationData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Registration default data",
                        "schema": {
                            "$ref": "#/definitions/controller.UserData"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/authentication/registration/verify": {
            "post": {
                "description": "Send user data and verification code to verify validity of verification.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Endpoint to verify user registration link",
                "parameters": [
                    {
                        "description": "The User Registration data and code",
                        "name": "RegistrationData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.VerifyRegistration"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Registration default data",
                        "schema": {
                            "$ref": "#/definitions/controller.RegistrationObject"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/token/reset": {
            "get": {
                "description": "Pass token in the Authorization header. Verify the token and send a new one with updated credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Reset token renews old token string",
                "responses": {
                    "200": {
                        "description": "Check validity",
                        "schema": {
                            "$ref": "#/definitions/controller.VerifyToken"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/token/user": {
            "get": {
                "description": "Pass token in the Authorization header. Verify token checks if the token is verified and send the authenticated user application related data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Checks if an authenticated user token is valid, and send user data of authenticated user",
                "responses": {
                    "200": {
                        "description": "Check validity",
                        "schema": {
                            "$ref": "#/definitions/controller.LoginObject"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/token/verify": {
            "get": {
                "description": "Pass token in the Authorization header. Verify token checks if the token is verified",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Checks if an authenticated user token is valid.",
                "responses": {
                    "200": {
                        "description": "Check validity",
                        "schema": {
                            "$ref": "#/definitions/controller.VerifyToken"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/user/": {
            "patch": {
                "description": "Update user data. This happens when a user wants to update user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Update user information",
                "parameters": [
                    {
                        "description": "Update user data json object",
                        "name": "UserData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UserData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User related data",
                        "schema": {
                            "$ref": "#/definitions/controller.UserData"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/apicore/user/{uid}": {
            "get": {
                "description": "Get user uid from url and retrieve user data with the uid string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Retrieve user data from user id",
                "responses": {
                    "200": {
                        "description": "User related data",
                        "schema": {
                            "$ref": "#/definitions/controller.UserData"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.ChangePassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "controller.LoginData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "controller.LoginObject": {
            "type": "object",
            "properties": {
                "token_string": {
                    "type": "string"
                },
                "user_data": {
                    "$ref": "#/definitions/controller.UserData"
                },
                "user_roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.UserRole"
                    }
                }
            }
        },
...
        "controller.RegistrationData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "controller.RegistrationObject": {
            "type": "object",
            "properties": {
                "token_string": {
                    "type": "string"
                },
                "user_data": {
                    "$ref": "#/definitions/controller.UserData"
                },
                "user_role": {
                    "$ref": "#/definitions/controller.UserRole"
                }
            }
        },
        "controller.ResponseObject": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.SocialMedia": {
            "type": "object",
            "properties": {
                "facebook": {
                    "type": "string"
                },
                "instagram": {
                    "type": "string"
                },
                "twitter": {
                    "type": "string"
                }
            }
        },
...
        "controller.UserData": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_picture": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "socialMedia": {
                    "$ref": "#/definitions/controller.SocialMedia"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "controller.UserRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "role_code": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "controller.VerifyRegistration": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "verification_code": {
                    "type": "string"
                }
            }
        },
        "controller.VerifyToken": {
            "type": "object",
            "properties": {
                "token_string": {
                    "type": "string"
                },
                "validity": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}